requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alembic==1.4.3            # via -r requirements.in
arrow==0.17.0             # via -r requirements.in
asyncpg==0.21.0           # via -r requirements.in
attrs==20.3.0             # via pytest
click==7.1.2              # via uvicorn
environs==9.2.0           # via -r requirements.in
fastapi==0.61.2           # via -r requirements.in
greenlet==0.4.17          # via sqlalchemy
gunicorn==20.0.4          # via -r requirements.in
h11==0.11.0               # via uvicorn
httptools==0.1.1          # via -r requirements.in
iniconfig==1.1.1          # via pytest
mako==1.1.3               # via alembic
markupsafe==1.1.1         # via mako
marshmallow==3.9.1        # via -r requirements.in, environs
packaging==20.7           # via pytest
pluggy==0.13.1            # via pytest
psycopg2==2.8.6           # via -r requirements.in
py==1.9.0                 # via pytest
pydantic==1.7.2           # via fastapi
pyjwt==1.7.1              # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest==6.1.2             # via -r requirements.in
python-dateutil==2.8.1    # via alembic, arrow
python-dotenv==0.15.0     # via environs
python-editor==1.0.4      # via alembic
python-magic==0.4.18      # via -r requirements.in
python-multipart==0.0.5   # via -r requirements.in
six==1.15.0               # via python-dateutil, python-multipart
sqlalchemy==1.4.0b1       # via -r requirements.in, alembic
starlette==0.13.6         # via fastapi
toml==0.10.2              # via pytest
uvicorn==0.12.3           # via -r requirements.in
uvloop==0.14.0            # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
